+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,6 +1,5 @@
-## 0.4.7+1
+## NEXT
 
-* Fixes StoreKit 2 purchase flow to send cancelled/pending/unverified results to `purchaseStream`.
 * Fixes Xcode 26.2 analyzer warnings in example app tests.
 
 ## 0.4.7